Update and add tests (WIP)
